Add gaming layer and media keys
Gaming layer is assigned to layer 3. The layer mostly moves around the modifiers
and space key to fit on the left half of the board so it can be used for games
which need those keys.

Media key support has been added.

I also moved switch around the function keys to the numbers/nav layer since I
think that makes more sense.